Make `riverdriver/riverpgxv5` its own submodule
Here, make `riverdriver/riverpgxv5` its own submodule. This does nothing
for the time being since the top-level River package is entirely
dependent on `riverpgxv5` as the only driver (and its used in examples
and tests), but the general idea is that in the future each driver could
bring in its own dependencies, and applications using River wouldn't
have to pull in every 3rd party database package that River supports.

The only problem is that there's a tradeoff: the way I had `riverpgxv5`
before is that it referenced `riverdriver.Driver` and made use of an
`internal.Sentinel` value from the top-level package to guarantee no one
could implement the driver interface externally. That no longer works
because it introduce a dependency cycle: `river` needs `riverpgxv5` and
`riverpgxv5` needs `river`. So I can remove the top-level types
referenced in `riverpgxv5` so the dependency only goes one way, but we
lose `internal.Sentinel`, and can no longer guarantee that `Driver`
isn't implemented.

So put simply: `riverpgxv5` gets its own module, but we move to only a
soft guarantee that external users don't implement `Driver` (we spell it
out in the docs, but there's nothing technical preventing people from
doing it).

Overall: I think it's probably worth it. It's pretty clear `Driver`
isn't meant for external implementation, there's no advantage to doing
so, and the interface is already a little leaky because users could
technically call functions on the interface already like `GetDBPool()`
and break themselves as those disappear in the future.