Documentation fix for the `UniqueOpts.ByState` parameter
A few weeks back we ended up changing the default set of states for the
`UniqueOpts.ByState` property -- previously, it hadn't included completed
jobs, but it was changed so that it did.

I thought I'd fixed the documentation as I was bringing that in, but
apparently I forgot. Here, update the Godoc for `UniqueOpts.ByState` to
also include completed jobs.