Bump comments in the unique job example
An incredibly trivial change to bump a few comments in the unique job
example in accordance to how they're written in our documentation. I
realized one of the comments implied that the job would be worked
exactly once which isn't a guarantee that we can actually make. Say that
insertion will be skipped instead.