Add names to variables in `Worker` interface (#14)
This is another small one: add names to variables in functions that are
in the `Worker` interface. These aren't strictly needed and may be
omitted, but are helpful in the case of IDEs like VScode so that when
users tell their IDE to implement an interface, the IDE generates
functions with nice variable names instead of stand-in boilerplate.