Refactor: Reduce rescuer LOCs / number of queries + optimize (#28)
I've been looking at ways we could start consolidating queries in
`river_job.sql` a bit because the file's currently quite a bit of a
mess. One pretty easy one that came out is that we multiple queries for
the rescuer that can be just one without much trouble.

Here, make that change, which also allows to simplify the rescuer's core
run loop by quite a bit and reduce the lines of code in there. Also
expect a slight performance improvement as we need only one round trip
to the database and the streamlined code lets us preallocate a number of
Go slices and reduce the number of `append`s happening.