Add the job timeout default to the godoc for `Config.JobTimeout`
A very small change to add the job timeout default in words to the
godoc for `Config.JobTimeout` so that it's easily visible in
documentation, and so that a user doesn't have to go chasing around for
the value in code.

There is some nominal doc rot risk in case we change this number, but
probably it's worth it for added convenience. We're also living in that
world already because ~all of the other configuration properties
document their defaults.