effectively disable reindexer (#34)
At this time, we're not sure we want to have a reindexer process
automatically running. It could be a net negative for many use cases and
we may want to do further work to avoid its potential downsides. For
now, disable it by simply removing all the indexes from its default
index list.