Comment out `replace` directive for `riverpgxv5`
I was trying to get CI working for the River demo project [1] and found
that it's not currently possible to install the River CLI to another Go
project:

    Run go install github.com/riverqueue/river/cmd/river@latest

    go: downloading github.com/riverqueue/river v0.0.4
    go: github.com/riverqueue/river/cmd/river@latest (in github.com/riverqueue/river@v0.0.4):
            The go.mod file for the module providing named packages contains one or
            more replace directives. It must not contain directives that would cause
            it to be interpreted differently than if it were the main module.
    Error: Process completed with exit code 1.

This is roughly consistent with what I'd previously understood for
`replace`. It's useful for debugging problems during development, but I
don't think it's meant to be a permanent decoration on a project. A
decent compromise is to keep it commented out, but available in case
it's needed to debug something, which is what I've done here.

Also, probably goes to show why it's important to keep a demo app
maintained so we can notice problems of this sort.

[1] https://github.com/bgentry/riverdemo/pull/7