Demonstrate internal job row type using struct conversion
I wanted to throw this one up as a possible alternative to a public
facing `rivertype` module in #36, in which we demonstrate an alternate
strategy using struct conversions. The problems it aims to solve are the
same:

* A way to access core job types from within subpackages, allowing more
  modularity (e.g. `Rescuer` could live with other maintenance services).

* Long term, a way of giving drivers access to job primitives without
  having to reference the top-level `river`.

The advantage of this approach is there's fewer user-facing changes, and
the user never has to interact with an alternate package.

It works by doing a struct conversion as necessary when moving from an
internal package to one that bubbles up to `river`. Go allows pure
struct conversions as long as their fields are the same, but there are
some limitations:

* Substructs like `AttemptError` can't be converted, even if identical.

* Custom types like `JobState` can't be converted, even if identical.

We work around these issues by:

* Keeping errors as a `[][]byte` instead of it being unmarshaled by pgx.
  A helper called `ErrorsUnmarshaled()` is added to the top-level
  `JobRow` to give easy access to unmarshaled errors.

* `State` is changed to a basic string akin to functions in `http`. This
  is technically a little less type-safe, but practically speaking would
  make very little difference because all these string-based types can
  be converted to and from each other with no checks anyway. Also, users
  aren't expected to access job state all that often. When they do, they
  can still use `job.State == river.JobStateAvailable` because the
  constants have been made available as strings.

I think this approach does also work reasonably well, although I worry a
little bit that there may be some sharp edge in the future that ends up
limiting us in some unexpected way. Also, having to use helpers like
`ErrorsUnmarshaled()` is definitely a little bit worse. That said, ~no
user-facing API changes, and has some marginal performance benefits:

* Converting structs between each other should be ~free.

* Errors probably aren't used that much, so by not unmarshaling them
  most of the time, we save on a lot of unnecessary JSON parsing.

Unlike #36, I didn't push the rescuer down (just in case this is a total
dead end), but this approach should make that possible.