Allow a `NULL` tags property to be `nil` when read from the database
This one's so tiny that it's minutiae, but as I was looking at the
null-ability of various job row fields today while implementing Ruby, I
realized that it doesn't really seem necessary to be marshaling a `NULL`
tags value to an empty `[]string{}` slice. In Go, `[]string{}` and
`[]string(nil)` are functionally identical in every way that I can think
of (`append` on `nil` works, `len` on `nil` works), so I think it makes
sense just to avoid the allocation and reflect what the database's
actual value is.

[1] https://go.dev/tour/moretypes/15