fix v003 migration: set default + update rows before set not null
The migration introduced in #53 / c5bd949 was flawed because it
attempted to set a default and also set not null on an existing field
all in one statement. This of course won't work if there are any
existing rows in the table, because those need to be updated to be
non-null before adding the not null constraint.

This change is made in place on the existing migration (rather than a
new one) because it fixes the fact that the migration simply can't
succeed on existing installs.

Fixes #63.