New `rivermigrate` Go API for running migrations from code + target version support
Here, as requested by several users, add a Go API that's able to run
migrations as an alternative to migrating via the CLI. We add a new
package `rivermigrate` to mirror the conventions of `riverdriver` and
`rivertest`.

`rivermigrate` is largely just a refactor of the existing internal
package `internal/dbmigrate` with a couple important tweaks:

* We give it the ability to take a driver the same way as we do for
  `river.Client` and `rivertest`.

* To mirror client and `rivertest` we add a transaction-specific variant
  that takes a `TTx` (`MigrateTx`).

* Because we now have the non-transaction and transaction variants, I
  take away the `Down`/`Up` function distinctions, and make up/down a
  parameter to `Migrate`/`MigrateTx` instead (so we only need two
  functions instead of four).

* I move to using an `embed.FS` instead of specific embedded strings for
  each version and the file system is walked to get all migrations
  within it. This will make adding new migrations in the future easier
  and more foolproof.

I also added a new option for `TargetVersion` so that it's possible to
do a limited migration to a specific version instead of going all the
way up or down. This is a pretty standard feature in most migration
frameworks.

This still can't support Goose because it takes an `sql.DB`, but I think
we can get there by adding a minimal driver for the `sql` package that
can run migrations and nothing else. More of this to be explored later.