More fixes for golangci-lint upgrade
So after #93 came in I ran my own `brew upgrade golangci-lint`, and I
guess I got a new point release, because I got a whole bunch of new lint
problems. I wish it was clearer which version I even got, but apparently
golangci-lint is releasing HEAD versions to Homebrew:

    $ golangci-lint --version
    golangci-lint has version HEAD-185b2c7 built with go1.21.4 from 185b2c7 on 2023-11-29T17:26:42Z

I fixed most of the problems, but I ended up disabling one of the
features of `testifylint` that disallows uses of testify functions in
goroutines, which would've required some very significant changes, and
which in our project produces a false positive rate (of actual problems)
of 100.0% as far as I can tell. See the description of this feature
here:

https://github.com/Antonboom/testifylint#go-require

    $ golangci-lint run --fix
    client_test.go:378:6: go-require: require must only be used in the goroutine running the test function (testifylint)
                                            require.NoError(t, err)
                                            ^
    internal/baseservice/base_service_test.go:40:4: go-require: require must only be used in the goroutine running the test function (testifylint)
                            require.FailNow(t, "Test case took too long to run (sleep statements should return immediately)")
                            ^
    internal/dbadapter/db_adapter_test.go:409:6: go-require: require must only be used in the goroutine running the test function (testifylint)
                                            require.NoError(t, err)
                                            ^
    internal/riverinternaltest/riverinternaltest_test.go:78:8: go-require: TestTx contains assertions that must only be used in the goroutine running the test function (testifylint)
                            _ = TestTx(ctx, t)